file Feature-class changed + Abstract class added
